+++ liminal/L.ts
@@ -10,8 +10,8 @@ export * from "./actions/enableTool.ts"
 export * from "./actions/fork.ts"
 export * from "./actions/getMessages.ts"
 export * from "./actions/infer.ts"
-export * from "./actions/pushLanguageModel.ts"
 export * from "./actions/setEmbeddingModel.ts"
+export * from "./actions/setLanguageModel.ts"
 export * from "./actions/setMessages.ts"
 export * from "./actions/throw.ts"
 export * from "./patterns/clear.ts"
